Refactor msvc no-vtable attribute to support macro
General:
refactor: replaced usage of declspec attribute to suppress vtable generation with support macro to enable multi-platform support

+++ src/Engine.Serialization/Layout/DataLayoutBase.cpp
@@ -8,84 +8,84 @@ using namespace hg::engine::serialization;
 using namespace hg;
 
 DataLayoutBase::DataLayoutBase(const bool hasReflectedClass_) :
-    DataLayoutReflect(),
-    DataLayoutDispatch(*this),
-    _hasReflectedClass(hasReflectedClass_) {}
+	DataLayoutReflect(),
+	DataLayoutDispatch(*this),
+	_hasReflectedClass(hasReflectedClass_) {}
 
 DataLayoutBase::~DataLayoutBase() {
 
-    tidy();
+	tidy();
 
-    DataLayoutReflect::~DataLayoutReflect();
-    DataLayoutDispatch::~DataLayoutDispatch();
+	DataLayoutReflect::~DataLayoutReflect();
+	DataLayoutDispatch::~DataLayoutDispatch();
 }
 
 void DataLayoutBase::tidy() {
 
-    for (auto& define : _defines) {
-        delete define;
-        define = nullptr;
-    }
+	for (auto& define : _defines) {
+		delete define;
+		define = nullptr;
+	}
 
-    _defines.clear();
+	_defines.clear();
 }
 
 cref<DataLayoutReflect> DataLayoutBase::reflect() const noexcept {
-    return static_cast<cref<DataLayoutReflect>>(*this);
+	return static_cast<cref<DataLayoutReflect>>(*this);
 }
 
 ref<DataLayoutReflect> DataLayoutBase::reflect() noexcept {
-    return static_cast<ref<DataLayoutReflect>>(*this);
+	return static_cast<ref<DataLayoutReflect>>(*this);
 }
 
 cref<DataLayoutDispatch> DataLayoutBase::dispatch() const noexcept {
-    return static_cast<cref<DataLayoutDispatch>>(*this);
+	return static_cast<cref<DataLayoutDispatch>>(*this);
 }
 
 bool DataLayoutBase::hasReflectedClass() const noexcept {
-    return _hasReflectedClass;
+	return _hasReflectedClass;
 }
 
 bool DataLayoutBase::hasDynamicDefines() const noexcept {
 
-    for (const auto& entry : _defines) {
-        if (entry->hasSubDefines()) {
-            return true;
-        }
-    }
+	for (const auto& entry : _defines) {
+		if (entry->hasSubDefines()) {
+			return true;
+		}
+	}
 
-    return false;
+	return false;
 }
 
 bool DataLayoutBase::justConstantDefines() const noexcept {
 
-    for (const auto& entry : _defines) {
-        if (entry->hasSubDefines()) {
-            return false;
-        }
-    }
+	for (const auto& entry : _defines) {
+		if (entry->hasSubDefines()) {
+			return false;
+		}
+	}
 
-    return true;
+	return true;
 }
 
 s64 DataLayoutBase::size() const noexcept {
 
-    s64 size { 0LL };
-    for (const auto& entry : _defines) {
+	s64 size { 0LL };
+	for (const auto& entry : _defines) {
 
-        const auto last { entry->_offset + entry->size() };
-        assert(last != entry->_offset);
+		const auto last { static_cast<s64>(entry->_offset) + entry->size() };
+		assert(last != entry->_offset);
 
-        size = MAX(size, last);
-    }
+		size = MAX(size, last);
+	}
 
-    return size;
+	return size;
 }
 
 cref<decltype(DataLayoutBase::_defines)> DataLayoutBase::defines() const noexcept {
-    return _defines;
+	return _defines;
 }
 
 void DataLayoutBase::define(mref<ptr<layout::LayoutDefine>> define_) {
-    _defines.push_back(std::move(define_));
+	_defines.push_back(std::move(define_));
 }
